build(typescript): add tsconfig.tsbuildinfo for incremental builds
Enable and check in the TypeScript build info file to leverage
incremental compilation. This speeds up subsequent builds by only
recompiling files that have changed since the last compilation.

This file was generated after setting the `composite` flag to true in
the tsconfig.json file, which is a prerequisite for project
references and optimized build workflows.